creation of the control algorithm
